do not disable heartbeat consumer during boot slave procedure
Instead of disabling the heartbeat consumer while a slave is booting, we simply
ignore heartbeat events. The problem with disabling the consumer is that one CAN
frame receiver (that detects boot-up events) stops another CAN frame receiver
(the heartbeat consumer) while both are processing the same message. This is
undefined behavior.