retry the first SDO during the NMT 'boot slave' process on timeout
The first SDO upload request during the NMT 'boot slave' process may occur
immediately after an NMT 'reset communication' command. Since the slave is
undergoing a state change, it is possible for it to miss the request. Retrying
the SDO on timeout gives the slave some extra time to respond, without
introducing a wait for fast slaves.