util: fix 64-bit value truncation on x86_64-w64-mingw32
This bug is a result of an incorrect assumption that 'long' data type in
MinGW-w64 has a size of 8 bytes. Actual 'long' size is 4 bytes on all Windows
versions and the same applies to MinGW-w64.

This lead to a 64-bit value truncation in clz64(), ctz64(), ffs64(),
parity64() and popcount64() functions.