rpdo.c: using `assert` instead of `if` for guaranteed condition.
It's now impossible for `co_rpdo_init_timer_swnd` to be called in
non-SYNC mode. It's only call-site is in `co_rpdo_sync` right after
an explicit check for SYNC mode.

Promoted related part of a conditional expression to an assert as it
was impossible to cover it in tests.