dev.c: removed undefined behavior in co_dev_write_dcf().
Now we only increment the pointer (now called bp for byte pointer)
only if it's not NULL to avoid UB. We also have a separate bytes
variable to count how many bytes were written / are required.

Additionally, calling `co_val_write()` with a basic type cannot fail,
so I removed returned value check. Also removed the first call as it
was not necessary - incrementing bp and bytes variables is enough to
reserve storage.