libc: Updated `LELY_HAVE_TIMESPEC` macro handling
Closes #102

This particular definition was being handled differently than
other definitions - specifically, it used `undef` instead of
`ifdef`. This makes it difficult to override this setting when
attempting to build on baremetal.

This change updates the handling to match the rest of the
code and allows the user to override the setting at
compile time.