Fix phantom branches in <lely/util/endian.h>.
The header file `endian.h` is now included in the `lely-unit-test.hpp`
header file. That in turn is included in `lely-unit-test.cpp` which
was compiled with C++ exception support enabled, which resulted in
additional branches generated by gcc and seen by gcov in functions
`ldle_u16` and `ldle_u32`. Only those 2 functions are used in libtest,
hence only those had phantom branches.

This commit adds `-fno-exceptions` to `libtest` so that it's
consistent with other test libraries.