Silence a GCC 11.1 warning.
Initialize a variable that's passed later as `const void*`. Compiler
assumes it won't be modified but will likely be read at some
point. That was enough to trigger a -Wmaybe-uninitialized warning that
apparently had some improvements made in the 11.x release.

This is the only issue that needs to be solved to support compilation
using gcc 11.1. It's been released over a month ago and ArchLinux has
already updated to that version over 2 weeks ago.