co: add calls to download indication function before every SDO response
This commit add extra calls to the download indication function before every
SDO response. In particular right after a non-expedited SDO request is received,
so that the callback can check the size, and right before a block confirmation
is sent, so the callback can verify and/or process the block.